+++ metadata.rb
@@ -1,12 +1,12 @@
-name             'exabgp'
-maintainer       'Aetrion, LLC.'
+name 'exabgp'
+maintainer 'Aetrion, LLC.'
 maintainer_email 'ops@dnsimple.com'
-license          'Apache 2.0'
-description      'Installs/Configures exabgp'
+license 'Apache 2.0'
+description 'Installs/Configures exabgp'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          IO.read(File.join(File.dirname(__FILE__), 'VERSION')) rescue "0.0.1"
 
 supports 'ubuntu', '>= 12.04'
 
+depends 'apt'
 depends 'python'
 depends 'runit'