+++ recipes/default.rb
@@ -3,6 +3,7 @@
 # Recipe:: default
 #
 # Copyright 2012, DNSimple, Inc.
+# Copyright 2015, Quadra rubra, UAB.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -17,13 +18,6 @@
 # limitations under the License.
 #
 
-include_recipe 'python'
-include_recipe 'runit'
-
-python_pip 'exabgp' do
-  action :install
-end unless node[:recipes].include? 'exabgp::source'
-
 directory '/etc/exabgp'
 
 template '/etc/exabgp/route_0_watchdog.sh' do
@@ -35,25 +29,21 @@ end
 template 'exabgp: config' do
   path '/etc/exabgp/exabgp.conf'
   source 'exabgp.conf.erb'
-  variables( router_id: node.ipaddress,
-             hold_time: node[:exabgp][:hold_time],
-             neighbor_ipv4: node[:exabgp][:ipv4][:neighbor],
-             local_address_ipv4: node.ipaddress,
-             route_ipv4: node[:exabgp][:ipv4][:anycast].gsub(/\d+$/, '0'),
-             prefix_ipv4: node[:exabgp][:ipv4][:prefix],
-             enable_ipv4_static_route: node[:exabgp][:ipv4][:enable_static_route],
+  variables(router_id: node.ipaddress,
+            hold_time: node[:exabgp][:hold_time],
+            neighbor_ipv4: node[:exabgp][:ipv4][:neighbor],
+            local_address_ipv4: node.ipaddress,
+            route_ipv4: node[:exabgp][:ipv4][:anycast].gsub(/\d+$/, '0'),
+            prefix_ipv4: node[:exabgp][:ipv4][:prefix],
+            enable_ipv4_static_route: node[:exabgp][:ipv4][:enable_static_route],
 
-             neighbor_ipv6: node[:exabgp][:ipv6][:neighbor],
-             local_address_ipv6: node[:ipv6address],
-             route_ipv6: node[:exabgp][:ipv6][:anycast].gsub(/\d+$/, ''),
-             prefix_ipv6: node[:exabgp][:ipv6][:prefix],
-             local_as: node[:exabgp][:local_as],
-             peer_as: node[:exabgp][:peer_as],
-             community: node[:exabgp][:community].join(' '))
+            neighbor_ipv6: node[:exabgp][:ipv6][:neighbor],
+            local_address_ipv6: node[:ipv6address],
+            route_ipv6: node[:exabgp][:ipv6][:anycast].gsub(/\d+$/, ''),
+            prefix_ipv6: node[:exabgp][:ipv6][:prefix],
+            local_as: node[:exabgp][:local_as],
+            peer_as: node[:exabgp][:peer_as],
+            community: node[:exabgp][:community].join(' '))
   mode '644'
   notifies :restart, 'service[exabgp]'
-end
-
-runit_service 'exabgp' do
-  default_logger true
 end